Improve property setting and configuration options
Manually setting every possible top-level attribute is far too
inflexible, especially given that this gem is already missing a number
of optional top-level keys. Having to add documentation (that's
relatively useless), a line in the initializer, lines in the
serializers, extra spec cases, etc, is a lot of work for what should be
low effort.

Also added a `Config` class to encapsulate some configuration options,
such as a `time_formatter` so that applications can choose how they want
to perform time formatting. For example, someone could write the
following formatter:

```ruby
def format_time(time)
  if time.is_a?(Integer)
    time
  elsif time.is_a?(Time)
    time.to_i * 1_000
  else
    raise ArgumentError, 'formatter accepts integers and Time objects'
  end
end
```

Similar formatters can be specified for `event_properties` and
`user_properties`, allowing the developer to choose what intermediary
format they want for those fields and how to convert that to a valid
form for sending to Amplitude.